Add files via upload
This project applies Naive Bayes and Logistic Regression Algorithms to a banknote dataset from UC Irvine's repository, The algorithms were implemented in Numpy and Pandas without the help of any other libraries. 3-fold cross validation is performed and the accuracy of the two algorithms is compared and plotted, using a different fraction of the training set each time.